create user

+++ modules/item/biz/delete_item.go
@@ -2,6 +2,7 @@ package biz
 
 import (
 	"context"
+	"errors"
 	"todolist/common"
 	"todolist/modules/item/model"
 )
@@ -12,11 +13,12 @@ type DeleteItemStore interface {
 }
 
 type deleteItemBiz struct {
-	store DeleteItemStore
+	store     DeleteItemStore
+	requester common.Requester
 }
 
-func NewDeleteItemBiz(store DeleteItemStore) *deleteItemBiz {
-	return &deleteItemBiz{store: store}
+func NewDeleteItemBiz(store DeleteItemStore, requester common.Requester) *deleteItemBiz {
+	return &deleteItemBiz{store: store, requester: requester}
 }
 
 func (biz *deleteItemBiz) DeleteItemBiz(ctx context.Context, id int) error {
@@ -26,6 +28,10 @@ func (biz *deleteItemBiz) DeleteItemBiz(ctx context.Context, id int) error {
 	}
 	if data.Status == "Deleted" {
 		return common.ErrEntityDeleted(model.EntityName, err)
+	}
+
+	if biz.requester.GetUserId() != data.UserId {
+		return common.ErrNoPermission(errors.New("No permission"))
 	}
 
 	if err := biz.store.DeleteItem(ctx, map[string]interface{}{"id": id}); err != nil {
